This is  is actually a search algorithm.
search begins with already sorted elements. x is compared with middle element where data is divided into left and right. If x matches with mid element it returns the mid value. If x is greater than middle element then x is shifted to right side where the process is repeated. Same goes for the left which is if mid value is less than x then x is shifted to left side and then compared with mid value of the left sub array.